The RoomDB
The RoomDB will serve as the system wide source of truth of active rooms.

Some goals of the Rooms DB subsystem

1.) We should be able to give out a unique room name. There should be no
collisions in the room names.

2.) We should be able to quickly answer the question "Is this room name
taken?"

ETS was chosen because another project wide goal is few (no?)
outside dependencies. ETS is a fast simple choice to holding server wide
state. Currently the plan is to only support a single server running at
a time, which means ETS being an in memory DB isn't that big of a deal.
In the future it could be switched out to a more distributed data store.