Add a uuid to Rooms
This uuid will serve as a unique room identifier, this means that if a
room name gets recycled, any persistent references will not spill over.

An example, if I'm authorized to view room "test", and the server gets
reset, its important that I'm no longer able to access the room because
it could be owned by someone else now. For any persistent records like
cookies, the ref value should be used.

I'm experiementing with using erlang references as guids instead of the
more traditionall 128 binary strings that you see most places, I'll see
if it makes it easier or not

(It would be a pretty cool contribution to ELixir to have a Guid Module
to deal with 128 bit references)